Hybris demo build

+++ Logic/source/AI/EntityManager.cpp
@@ -15,7 +15,7 @@ using namespace Logic;
 #include <Player\Player.h>
 #include <Projectile\ProjectileManager.h>
 
-#include <Graphics\include\Renderer.h>
+
 #include <Graphics\include\Structs.h>
 #include <Physics\Physics.h>
 #include <Graphics\include\Particles\ParticleSystem.h>
@@ -50,6 +50,11 @@ EntityManager::~EntityManager()
     delete m_threadHandler;
 }
 
+void EntityManager::resetTriggers()
+{
+    m_triggerManager.reset();
+}
+
 void EntityManager::allocateData()
 {
     m_enemies.resize(AStar::singleton().getNrOfPolygons());
@@ -59,7 +64,7 @@ void EntityManager::allocateData()
 void EntityManager::loadDebugCmds()
 {
 #ifdef _DEBUG
-    DebugWindow::getInstance()->registerCommand("SETAI", [&](std::vector<std::string> &para) -> std::string {
+    DebugWindow::getInstance()->registerCommand("LOG_SETAI", [&](std::vector<std::string> &para) -> std::string {
         try {
             m_aiType = static_cast<AIType> (stoi(para[0]));
             return "AI Mode Updated";
@@ -70,26 +75,46 @@ void EntityManager::loadDebugCmds()
 #endif
 }
 
-void EntityManager::deallocateData()
+void EntityManager::deallocateData(bool forceDestroy)
 {
     for (std::vector<Enemy*>& list : m_enemies)
     {
-        for (Enemy *enemy : list)
+        for (Enemy *&enemy : list)
         {
-            DeleteBody(*enemy); 
-            delete enemy;
-            enemy = nullptr;
+            if (forceDestroy || !enemy->hasCallbackEntities())
+            {
+                if (forceDestroy)
+                    enemy->clearCallbacks(false);
+                DeleteBody(*enemy);
+                delete enemy;
+                enemy = nullptr;
+            }
+            else {
+                m_deadEnemies.push_back(enemy);
+            }
         }
         list.clear();
     }
 
-    for (Enemy *enemy : m_deadEnemies)
+    // clear out enemies without callbacks in needs to do
+    for (Enemy *&enemy : m_deadEnemies)
     {
-        DeleteBody(*enemy);
-        delete enemy;
+        if (forceDestroy || !enemy->hasCallbackEntities()) 
+        {
+            if (forceDestroy)
+                enemy->clearCallbacks(false);
+            delete enemy;
+            enemy = nullptr;
+        }
     }
 
+    // copy over old list without destroyed enemies
+    std::vector<Enemy*> oldList = m_deadEnemies;
     m_deadEnemies.clear();
+    for (Enemy *enemy : oldList)
+        if (enemy)
+            m_deadEnemies.push_back(enemy);
+
     m_aliveEnemies = 0;
 }
 
@@ -107,7 +132,7 @@ void EntityManager::update(Player const &player, float deltaTime)
             {
                 updateEnemies(static_cast<int> (i), player, deltaTime);
                 if ((i + m_frame) % ENEMIES_PATH_UPDATE_PER_FRAME == 0)
-                    m_threadHandler->addWork({ this, static_cast<int> (i), player });
+                    m_threadHandler->addWork({ this, static_cast<int> (i), &player });
             }
         }
     }
@@ -115,11 +140,12 @@ void EntityManager::update(Player const &player, float deltaTime)
 
 	for (int i = 0; i < m_deadEnemies.size(); ++i)
 	{
-		m_deadEnemies[i]->updateDead(deltaTime);
+	    //m_deadEnemies[i]->updateDead(deltaTime); -- Delete body right now, maybe later keep the dead bodies, but it is not really important but can be changed
 	}
 
 	m_triggerManager.update(deltaTime);
 }
+
 void EntityManager::updateEnemies(int index, Player const &player, float deltaTime)
 {
 	bool goalNodeChanged = false;
@@ -156,13 +182,9 @@ void EntityManager::updateEnemy(Enemy *enemy, std::vector<Enemy*> &flock,
             flock[flock.size() - 1]
         );
 
+        DeleteBody(*enemy);
         m_deadEnemies.push_back(enemy);
         flock.pop_back();
-
-        // nice little death stuff
-        btVector3 oldVel = enemy->getRigidBody()->getLinearVelocity();
-        enemy->getRigidBody()->setLinearVelocity(btVector3());
-        enemy->getRigidBody()->applyCentralForce(oldVel * 1000);
     }
 }
 
@@ -211,13 +233,13 @@ Enemy* EntityManager::spawnEnemy(ENEMY_TYPE id, btVector3 const &pos,
     switch (id)
     {
     case ENEMY_TYPE::NECROMANCER:
-        enemy = newd EnemyNecromancer(Graphics::ModelID::ENEMYGRUNT, testBody, { 0.5f, 0.5f, 0.5f });
+        enemy = newd EnemyNecromancer(testBody, { 0.5f, 0.5f, 0.5f });
         break;
     case ENEMY_TYPE::NECROMANCER_MINION:
         enemy = newd EnemyChaser(testBody);
         break;
     default:
-        enemy = newd EnemyTest(Graphics::ModelID::ENEMYGRUNT, testBody, { 0.5f, 0.5f, 0.5f });
+        enemy = newd EnemyTest(testBody, { 0.5f, 0.5f, 0.5f });
 		break;
     }
 
@@ -250,14 +272,14 @@ Trigger* EntityManager::spawnTrigger(int id, btVector3 const &pos,
     switch (id)
     {
     case 1: // wtf, starts at 1
-        trigger = m_triggerManager.addTrigger(Graphics::ModelID::JUMPPAD,
+        trigger = m_triggerManager.addTrigger(Resources::Models::UnitCube,
             Cube(pos, { 0, 0, 0 }, { 2, 0.1f, 2 }),
             500.f, physics, {},
             effectsIds,
             true);
         break;
     case 2:
-        trigger = m_triggerManager.addTrigger(Graphics::ModelID::AMMOBOX,
+        trigger = m_triggerManager.addTrigger(Resources::Models::UnitCube,
             Cube(pos, { 0, 0, 0 }, { 1, 0.5, 1 }), 0.f, physics, {},
             effectsIds,
             false);
@@ -286,15 +308,15 @@ int EntityManager::giveEffectToAllEnemies(StatusManager::EFFECT_ID id)
     return i;
 }
 
-void EntityManager::render(Graphics::Renderer &renderer)
+void EntityManager::render() const
 {
     for (int i = 0; i < m_enemies.size(); ++i)
     {
         for (Enemy *enemy : m_enemies[i])
         {
-            enemy->render(renderer);
+            enemy->render();
 #ifdef DEBUG_PATH	
-            enemy->debugRendering(renderer);
+            enemy->debugRendering();
 #endif
         }
     }
@@ -302,14 +324,14 @@ void EntityManager::render(Graphics::Renderer &renderer)
     for (int i = 0; i < m_deadEnemies.size(); ++i)
     {
 #ifndef DISABLE_RENDERING_DEAD_ENEMIES
-        m_deadEnemies[i]->render(renderer);
+     //   m_deadEnemies[i]->render();
 #endif
     }
 
-    m_triggerManager.render(renderer);
+    m_triggerManager.render();
 
 #ifdef DEBUG_ASTAR
-    AStar::singleton().renderNavigationMesh(renderer);
+    AStar::singleton().renderNavigationMesh();
 #endif
 }
